requirements.txt
requests
Flask==1.1.2
bs4
python-dotenv

